% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agronomic-norms.R
\name{agronomic_norms_fields}
\alias{agronomic_norms_fields}
\title{agronomic_norms_fields}
\usage{
agronomic_norms_fields(
  field_id,
  month_day_start,
  month_day_end,
  year_start,
  year_end,
  propertiesToInclude = "",
  exclude_years = NULL,
  accumulation_start_date = "",
  gdd_method = "standard",
  gdd_base_temp = 10,
  gdd_min_boundary = 10,
  gdd_max_boundary = 30,
  includeFeb29thData = TRUE,
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token,
  apiAddressToUse = awhereEnv75247$apiAddress
)
}
\arguments{
\item{-}{apiAddressToUse: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
dataframe of requested data for dates requested
}
\description{
\code{agronomic_norms_fields} pulls agronomic norm data from aWhere's API based on field id
}
\details{
This function allows you to calculate the averages for agronomic attributes
across any range of years for which data are available.  The data pulled includes norms for
growing degree days (GDDs), potential evapotranspiration (PET), Precipitation over
potential evapotranspiration (P/PET), accumulated GDDs, accumulated precipitation,
accumulated PET, and accumulated P/PET, along with the standard deviations
for these variables.  The data pulled is for the field id identified.
Default units are returned by the API.

The data returned in this function
allow you to compare this year or previous years to the long-term normals, calculated as
the average of those agronomic conditions on that day in that location over the years specified.

Note about dates: The system does not adjust for any difference in dates between the location of the user
          and where data is being requested from.  It is the responsibility of the user to ensure a valid
          date range is specified given any differences in timezone.  These differences can have implications
          for whether a given date should be requested from the daily_observed functions or the forecast functions
}
\examples{
\dontrun{agronomic_norms_fields(field_id = 'field_test'
                                ,month_day_start = '07-01'
                                ,month_day_end = '07-10'
                                ,year_start = 2008
                                ,year_end = 2016
                                ,exclude_years = "2010"
                                ,accumulation_start_date = ''
                                ,gdd_method = 'standard'
                                ,gdd_base_temp = 10
                                ,gdd_min_boundary = 10
                                ,gdd_max_boundary = 30)}
}
\references{
https://docs.awhere.com/knowledge-base-docs/historical-agronomic-norms/
}
