% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agronomic-values.R
\name{agronomic_values_area}
\alias{agronomic_values_area}
\title{agronomic_values_area}
\usage{
agronomic_values_area(
  polygon,
  day_start,
  day_end,
  propertiesToInclude = "",
  accumulation_start_date = "",
  gdd_method = "standard",
  gdd_base_temp = 10,
  gdd_min_boundary = 10,
  gdd_max_boundary = 30,
  numcores = 2,
  bypassNumCallCheck = FALSE,
  returnSpatialData = FALSE,
  verbose = TRUE,
  maxTryCount = 3,
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token,
  apiAddressToUse = awhereEnv75247$apiAddress
)
}
\arguments{
\item{-}{apiAddressToUse: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
data.frame of requested data for dates requested
}
\description{
\code{agronomic_values_area} pulls agronomic data from aWhere's API based on a data.frame of lat/lon, polygon or extent
}
\details{
This function returns agronomic data on GDDs, potential evapotranspiration (PET), Precipitation over
potential evapotranspiration (P/PET), accumulated GDDs, accumulated precipitation, accumulated PET, and
accumulated P/PET.  Default units are returned by the API.

Agronomic Values are calculated numbers that can be used to show the agronomic status of a field or crop.
These figures can be used, for example, to track and predict plant growth or identify water stress.
Accumulated values allow growers to easily identify how the weather has been over the season.
Both sets of data are commonly used on small and large farms alike.  This is a very flexible API
that supports a wide variety of configurations to get exactly the data you want as efficiently as
possible.

The polygon should be either a SpatialPolygons object or a well-known text character string or an extent.

Note about dates: The system does not adjust for any difference in dates between the location of the user
          and where data is being requested from.  It is the responsibility of the user to ensure a valid
          date range is specified given any differences in timezone.  These differences can have implications
          for whether a given date should be requested from the daily_observed functions or the forecast functions.
          Furthermore, because this function can take as input locations that may be in different timezones, it is
          the responsibility of the user to either ensure that the date range specified is valid for all relevant
          locations or to break the query into pieces.
}
\examples{
\dontrun{agronomic_values_area(polygon = raster::getData('GADM', country = "Gambia", level = 0, download = T),
                               ,day_start = '2018-04-28'
                               ,day_end = '2018-05-01'
                               ,numcores = 2)}
}
\references{
https://docs.awhere.com/knowledge-base-docs/agronomic-values-accumulations-by-geolocation/
}
