% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agronomic-values.R
\name{agronomic_values_fields}
\alias{agronomic_values_fields}
\title{agronomic_values_fields}
\usage{
agronomic_values_fields(
  field_id,
  day_start,
  day_end,
  propertiesToInclude = "",
  accumulation_start_date = "",
  gdd_method = "standard",
  gdd_base_temp = 10,
  gdd_min_boundary = 10,
  gdd_max_boundary = 30,
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token,
  apiAddressToUse = awhereEnv75247$apiAddress
)
}
\arguments{
\item{-}{apiAddressToUse: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
dataframe of requested data for dates requested
}
\description{
\code{agronomic_values_fields} pulls agronomic data from aWhere's API based on field id
}
\details{
This function returns agronomic data on growing degree days (GDDs), potential evapotranspiration (PET), Precipitation over
potential evapotranspiration (P/PET), accumulated GDDs, accumulated precipitation, accumulated PET, and
accumulated P/PET.  Default units are returned by the API.

Agronomic Values are calculated numbers that can be used to show the agronomic status of a field or crop.
These figures can be used, for example, to track and predict plant growth or identify water stress.
Accumulated values allow growers to easily identify how the weather has been over the season.
Both sets of data are commonly used on small and large farms alike.  This is a very flexible API
that supports a wide variety of configurations to get exactly the data you want as efficiently as
possible. It's also designed to work with the Fields and Plantings system to reduce the amount of input.
While a planting is not required to use this API, creating a Planting for your Fields will allow you
to get the most out of the aWhere platform.

Note about dates: The system does not adjust for any difference in dates between the location of the user
          and where data is being requested from.  It is the responsibility of the user to ensure a valid
          date range is specified given any differences in timezone.  These differences can have implications
          for whether a given date should be requested from the daily_observed functions or the forecast functions
}
\examples{
\dontrun{agronomic_values_fields(field_id = "field_test"
                                 ,day_start = "2018-11-01"
                                 ,day_end = "2018-11-30"
                                 ,gdd_method = "modifiedstandard"
                                 ,gdd_base_temp = 10
                                 ,gdd_min_boundary = 10
                                 ,gdd_max_boundary = 30)}
}
\references{
https://docs.awhere.com/knowledge-base-docs/agronomic-values-accumulations/
}
