% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkAccumulationStartDate}
\alias{checkAccumulationStartDate}
\title{Check Accumulation Start Date}
\usage{
checkAccumulationStartDate(accumulation_start_date, month_day_start)
}
\arguments{
\item{-}{month_day_start}
}
\description{
\code{checkAccumulationStartDate} Checks to see if the accumulation start date is valid
}
