% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkCredentials}
\alias{checkCredentials}
\title{Check Valid Credentials Loaded}
\usage{
checkCredentials()
}
\description{
\code{checkCredentials} Checks to see if valid aWhere API credentials are loaded
}

