% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataReturn.R
\name{checkDataReturn_forecasts}
\alias{checkDataReturn_forecasts}
\title{checkDataReturn_forecasts}
\usage{
checkDataReturn_forecasts(dataset, day_start, day_end, block_size)
}
\arguments{
\item{-}{dataset: the data.frame to check}

\item{-}{day_start: character string of the first day for which data was retrieved, in the form: YYYY-MM-DD}

\item{-}{day_end: character string of the last day for which data was retrieved, in the form: YYYY-MM-DD}

\item{-}{block_size: Integer value that corresponds to the number of hours to include in each time block.
Defaults to a 1 hour block.  This value must divide evenly into 24.}
}
\description{
\code{checkDataReturn_forecasts} Checks to see if data return is complete from forecasts API endpoint
}
