% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataReturn.R
\name{checkDataReturn_norms}
\alias{checkDataReturn_norms}
\title{checkDataReturn_norms}
\usage{
checkDataReturn_norms(
  dataset,
  monthday_start,
  monthday_end,
  year_start,
  year_end,
  exclude_year,
  includeFeb29thData
)
}
\arguments{
\item{-}{includeFeb29thData: Whether to keep data from Feb 29th on leap years.}
}
\description{
\code{checkDataReturn_norms} Checks to see if data return is complete from norms API
}
