% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkForecastSources}
\alias{checkForecastSources}
\title{Check Forecast Sources}
\usage{
checkForecastSources(sources)
}
\description{
\code{checkForecastSources} Checks that the requested Forecast source is valid
}
