% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkNormsStartEndDates}
\alias{checkNormsStartEndDates}
\title{Checks Start/End Dates for Norms}
\usage{
checkNormsStartEndDates(month_day_start, month_day_end)
}
\arguments{
\item{-}{monthday_start: character string of the month and day for the start
of the range of days you are calculating norms for, e.g., '07-01' (July 1)}

\item{-}{monthday_end: character string of the month and day for the end of the
range of days you are calculating norms for, e.g., '07-10' (July 10)}
}
\description{
\code{checkNormsStartEndDates} Checks Start/End Dates for Norms endpoint are valid
}
