% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkPropertiesEndpoint}
\alias{checkPropertiesEndpoint}
\title{Check Properties for Weather Endpoint}
\usage{
checkPropertiesEndpoint(endpoint, propertiesToInclude)
}
\arguments{
\item{-}{propertiesToInclude: vector of properties requested from API}
}
\description{
\code{checkPropertiesWeather} Check that the Years to be Requested for Norms is valid
}
