% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStatusCodeOfReturn.R
\name{checkStatusCode}
\alias{checkStatusCode}
\title{Check Status Code of Return}
\usage{
checkStatusCode(request)
}
\arguments{
\item{-}{request: object returned from HTTR}
}
\description{
\code{checkStatusCode} Checks to see if valid aWhere API credentials are loaded
}
