% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkValidField}
\alias{checkValidField}
\title{Check Valid Field ID}
\usage{
checkValidField(field_id, keyToUse, secretToUse, tokenToUse)
}
\arguments{
\item{-}{tokenToUse: aWhere API token to use}
}
\description{
\code{checkValidField} Checks to see if the Field ID is valid
}
