% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkValidLatLong}
\alias{checkValidLatLong}
\title{Check Valid Lat/Long}
\usage{
checkValidLatLong(latitude, longitude)
}
\arguments{
\item{-}{latitude: the latitude for the location for which you want data}

\item{-}{longitude: the latitude for the location for which you want data}
}
\description{
\code{checkValidLatLong} Checks to see if Lat/Long coordinates are valid
}
