% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkValidStartEndDatesForecast}
\alias{checkValidStartEndDatesForecast}
\title{Check Valid Start End Dates for Forecast}
\usage{
checkValidStartEndDatesForecast(day_start, day_end)
}
\arguments{
\item{-}{day_start: character string of start date in form: YYYY-MM-DD}

\item{-}{day_end: character string of end date in form: YYYY-MM-DD}
}
\description{
\code{checkValidStartEndDatesForecast} Checks to see if Start/End Dates passed are valid for Forecast
}
