% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-awhere-grid.R
\name{create_awhere_grid}
\alias{create_awhere_grid}
\title{Create aWhere Grid}
\usage{
create_awhere_grid(polygon)
}
\arguments{
\item{-}{polygon: either a SpatialPolygons data type object or a character string of valid WKT}
}
\description{
\code{create_awhere_grid} based on a SpatialPolygon, creates a data frame based on aWhere's grid system
}
\details{
Creates a grid of lat/lon points within given polygon.
The aWhere grid is spaced at .08333 decimal degrees resolution (or 5 arc-minutes),
so spacing a new grid at every .08 degrees should guarantee
a grid point in each aWhere grid cell.
}
