% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_job}
\alias{create_job}
\title{Create Job}
\usage{
create_job(
  api_requests,
  request_titles,
  job_title,
  job_type = "batch",
  verbose = TRUE,
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token
)
}
\arguments{
\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
- job_id: The Job ID. You will need this to retrieve the job status and results.
}
\description{
\code{create_job} This API will register a batch job in the aWhere platform.
}
\details{
Batch jobs allow you to execute many API calls in a single batch, which is much more efficient and faster than making hundreds
or thousands of individual requests. When you create a batch, you define each endpoint you want to call, and then aWhere's platform
steps through each internally and provides all the results at once.  Any GET request from any of our Version-2 APIs can be included

Batch jobs are queued and executed in the order they are created, from across all customers using aWhere APIs. While many jobs can run
concurrently, the rest are inserted into a queue and executed quite quickly. Use the Status and Results endpoint to monitor your job status
and retrieve the results when complete.

Important: This API will return a job ID which you must retain; there is not currently an API to list all jobs that you've created.
Important: There is a limit of 10,000 requests per batch job.
Note: The current conditions API cannot be included in batch jobs.
}
\examples{
\dontrun{create_job(c("GET /v2/weather/fields/field_test/observations", "GET /v2/weather/fields/aWhereOffice/observations"), c("farmA", "farmB"), "job_1")}
}
\references{
https://developer.awhere.com/api/reference/batch/create
}
