% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_planting}
\alias{create_planting}
\title{Create Planting}
\usage{
create_planting(field_id, crop, planting_date, proj_yield_amount = "",
  proj_yield_units = "", proj_harvest_date = "", yield_amount = "",
  yield_units = "", harvest_date = "", keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{field_id: an ID of your choosing (string)}

\item{-}{crop: cropId or crop name (string)}

\item{-}{planting_date: date crop was planted in the field. Format as YYYY-MM-DD (string)}

\item{-}{proj_yield_amount: amount of projected yield from planting (string)}

\item{-}{proj_yield_units: units of projected yield (string - optional)}

\item{-}{proj_harvest_date: projected harvest date at the start of the season. Format as YYYY-MM-DD (string - optional)}

\item{-}{yield_amount: actual yield (string - optional)}

\item{-}{yield_units: units of actual yield (string - optional)}

\item{-}{harvest_date: actual harvest date at end of season. Format as YYYY-MM-DD (string - optional)}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
- system generated planting id along with a print text that informs if the query succeeded or not
}
\description{
\code{create_planting} creates a planting in a field location in the aWhere platform for which you can request weather
}
\details{
Fields are how you manage the locations for which you're tracking weather, agronomics,
models, and progress over growing seasons. By registering a field, you create a quick way
to consistently reference locations across all our APIs, and allow our modeling APIs
to better operate and tune to the conditions and status of each specific field. A Planting
is the record of a crop's season in a given field, and is where you tell the platform
about what is planted there and when it was planted.

Creating a planting will provide the aWhere platform the information needed to run models
and more efficiently calculate agronomic values. You can also use these properties to record
projections for the field, like yield or harvest date, to track the success of a field over
the course of a growing season. Recording projected and actual yield and harvest date also helps
aWhere tune the models for even greater accuracy.

There can only be one active planting per field. You can create multiple plantings per field
but only the most recent one will be considered the "current" one. Use this functionality to
create historical records if you have them.

When creating a planting, you must specify the crop and planting date.
The crop must be an option from the Crops API; but there is also a short cut where if you don't
know or want to use a specific crop ID, you can simply specify the crop name, such as "corn" or
"wheat" and the the API will select the default for that category.

This script creates a planting in a field location in the aWhere platform. By setting an Id you can retrieve the weather
and agronomics for that location in all the other APIs. The planting ID corresponds to a planting within a field.
}
\examples{
\dontrun{create_planting(field_id='field_test',crop='corn', planting_date='2015-10-25', proj_yield_amount='100',
                         proj_yield_units='Bushels', proj_harvest_date='2016-02-01', yield_amount='110',
                         yield_units='Bushels', harvest_date='2016-02-01')
}
}
\references{
http://developer.awhere.com/api/reference/plantings/create
}
