% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-shapewkt.R
\name{create_shapewkt}
\alias{create_shapewkt}
\title{create_shapewkt}
\usage{
create_shapewkt(longitude, latitude)
}
\arguments{
\item{-}{longitude: longitude coordinate for centroid}

\item{-}{latitude: latitude coordinate for centroid}
}
\description{
\code{create_shapewkt} writes the WKT for a grid cell, based on a centroid coordinate
}
