% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createField.R
\name{CreateField}
\alias{CreateField}
\title{CreateField}
\usage{
CreateField(fieldId, latitude, longitude, farmId, fieldName = "",
  acres = "")
}
\arguments{
\item{-}{fieldId: an ID of your choosing (string)}

\item{-}{latitude: the latitude of the field location in decimal format. (string)}

\item{-}{longitude: the longitude of the field location in decimal format (string)}

\item{-}{farmId: an arbitrary ID for the farm to which this field belongs (string)}

\item{-}{fieldName: a name of the location (optional -string)}

\item{-}{acres: the acres of the field (optional -string)}
}
\value{
- printed text that informs if the query succeded or not
}
\description{
\code{CreateField} This API will register a field location in the aWhere platform. This is a one-time operation for each field.
}
\details{
Fields are how you manage the locations for which you're tracking weather, agronomics,
models, and progress over growing seasons. By registering a field, you create a quick way
to consistently reference locations across all our APIs, and allow our modeling APIs
to better operate and tune to the conditions and status of each specific field. A Planting
is the record of a crop's season in a given field, and is where you tell the platform
about what is planted there and when it was planted.

Creating a field registers the location with the aWhere system, making it easier to reference
and track your locations as well as run agronomics and models automatically. You
only need to create a field once, after which point you can reference it by the ID you create
(you'll use this ID in virtually every URI endpoint in our system).

All Spaces will be converted to underscores to conform with the requirements of the API
}
\examples{
CreateField("field123","39.8282","-98.5795","farmA","Some Field Location","100")
CreateField("field456","40.8282","-100.5795","farmA","Some Field Location","100")
}
\references{
http://developer.awhere.com/api/reference/fields/create-field
}

