% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-shapewkt.R
\name{custom_fortify}
\alias{custom_fortify}
\title{custom_fortify}
\usage{
custom_fortify(shapewkt)
}
\arguments{
\item{-}{shapewkt: a well-known text string for a grid cell polygon}
}
\description{
\code{custom_fortify} breaks up a shapewkt string into a numeric vector containing the points in the WKT
}
