% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-daily.R
\name{daily_observed_area}
\alias{daily_observed_area}
\title{daily_observed_area}
\usage{
daily_observed_area(polygon, day_start, day_end,
  propertiesToInclude = "", numcores = 2,
  keyToUse = awhereEnv75247$uid, secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{polygon: either a SpatialPolygons object or a well-known text string}

\item{-}{day_start: character string of the first day for which you want to retrieve data, in the form: YYYY-MM-DD}

\item{-}{day_end: character string of the last day for which you want to retrieve data, in the form: YYYY-MM-DD}

\item{-}{propertiesToInclude: character vector of properties to retrieve from API.  Valid values are temperatures, precipitation, solar, relativeHumidity, wind (optional)}

\item{-}{numcores: number of cores to use in parallel loop. To check number of available cores: parallel::detectCores()}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
data.frame of requested data for dates requested
}
\description{
\code{daily_observed_area} pulls historical weather data from aWhere's API for a provided polygon
}
\details{
This function returns weather data on Min/Max Temperature, Precipitation,
Min/Max Humidity, Solar Radiation, and Maximum Wind Speed,
Morning Max Windspeed, and Average Windspeed for the polygon passed to the function.
Default units are returned by the API. The polygon should be either a SpatialPolygons object or
a well-known text character string.

The Weather APIs provide access to aWhere's agriculture-specific Weather Terrain system,
and allows retrieval and integration of data across all different time ranges, long term normals,
daily observed, current weather, and forecasts. These APIs are designed for efficiency,
allowing you to customize the responses to return just the attributes you need.

Understanding the recent and long-term daily weather is critical for making in-season decisions.
This API opens the weather attributes that matter most to agriculture.
}
\examples{
\dontrun{daily_observed_area(polygon = raster::getData('GADM', country = "Gambia", level = 0, download = F),
                               ,day_start = '2018-04-28'
                               ,day_end = '2018-05-01')}
}
\references{
http://developer.awhere.com/api/reference/weather/observations/geolocation
}
