% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts.R
\name{forecasts_fields}
\alias{forecasts_fields}
\title{forecasts_fields}
\usage{
forecasts_fields(
  field_id,
  day_start = as.character(Sys.Date()),
  day_end = "",
  block_size = 1,
  useLocalTime = TRUE,
  returnOnlySoilVars = FALSE,
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token,
  apiAddressToUse = awhereEnv75247$apiAddress
)
}
\arguments{
\item{-}{apiAddressToUse: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
data.frame of requested data for dates requested
}
\description{
\code{forecasts_fields} pulls forecast weather data from aWhere's API based on field id
}
\details{
This function returns today's forecast plus the forecast for up to 7 more days. Forecasts are available
in many sizes of time blocks, from hourly to daily. The data this function returns is
Min/Max Temperature, Precipitation Amount, Chance of Precipitation,
Max Humidity, Relative Humidity, Solar Radiation, Average Wind Speed, Max Windspeed, Percentage
of Sky Covered by Clouds, and Percentage of Clear Sky for the field id specified.
Default units are returned by the API.

Note that when block_size = 1 the fields min/max relative humidity and min/max wind will be NA

The Weather APIs provide access to aWhere's agriculture-specific Weather Terrain system,
and allows retrieval and integration of data across all different time ranges long term normals,
daily observed, current weather, and forecasts. These APIs are designed for efficiency,
allowing you to customize the responses to return just the attributes you need.

Note about dates: The system does not adjust for any difference in dates between the location of the user
          and where data is being requested from.  It is the responsibility of the user to ensure a valid
          date range is specified given any differences in timezone.  These differences can have implications
          for whether a given date should be requested from the daily_observed functions or the forecast functions
}
\examples{
\dontrun{forecasts_fields(field_id = 'field_test'
                          ,day_start = as.character(Sys.Date())
                          , block_size = 12)
         forecasts_fields('field_test'
                          ,day_start = as.character(Sys.Date())
                          ,day_end = as.character(Sys.Date() + 5))}
}
\references{
https://docs.awhere.com/knowledge-base-docs/forecast-weather/
}
