% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agronomic-crops.R
\name{get_crops}
\alias{get_crops}
\title{Get Crops}
\usage{
get_crops(crop_id = "", keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{crop_id: Either a crop id to retrieve information for that specific crop
or an empty string to retrieve information on all crops associated
with the user's aWhere API account (string - optional)}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
- data.frame containing information about requested crop(s)
}
\description{
\code{get_crops} Calls Crops endpoint of API
}
\details{
Models are designed for specific crops, and often specific types or varieties of crops.
This API provides the list of available crops. When creating a planting record, you're encouraged
to specify the crop that is planted in the field using these records.

Within each category of crop (organized by name), there is a default crop. When creating a planting,
you may not know or care about the specific variety, and can simply specify the crop name as
what is planted there. When you do this, the system will use the default crop for that category.
You'll know which crop record is the default by referencing the isDefaultForCrop property, described below.
}
\examples{
\dontrun{get_crops('barley-generic')
         get_crops()
}
}
\references{
https://developer.awhere.com/api/reference/crops
}
