% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_model_details}
\alias{get_model_details}
\title{Get Model Details}
\usage{
get_model_details(model_id, keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{model_id: A model id to retrieve details for that specific model}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
- data.frame containing details about requested model
}
\description{
\code{get_model_details} Calls Model Details endpoint of API
}
\details{
Every model has its own parameters and criteria it uses to calculate the results.
This API will return those details; for the growth stage models currently available
this API returns information about the crop stages that the model may return.
}
\examples{
\dontrun{get_model_details('BarleyGenericMSU')
}
}
\references{
https://developer.awhere.com/api/reference/models/details
}
