% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_model_results}
\alias{get_model_results}
\title{Get Model Results}
\usage{
get_model_results(field_id, model_id, keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{model_id: A model id to retrieve details for that specific model}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
- data.frame containing details about requested model
}
\description{
\code{get_model_details} Calls Model Results endpoint of API
}
\details{
Running a model is as simple as requesting its current results. The aWhere platform
will use the information previously provided about a field and planting to generate
the relevant, field-specific information. For growth stage models, the results are
the current stage as well as information about the most recent and next stage.

Important: Models require the use of the Fields and Plantings APIs to store important
information about a field and crop. When a Field is specified for this API, it will
select the most recently created Planting record to retrieve the crop and planting date.
}
\examples{
\dontrun{get_model_results('Field1', 'BarleyGenericMSU')
}
}
\references{
https://developer.awhere.com/api/reference/models/results
}
