% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_models}
\alias{get_models}
\title{Get Models}
\usage{
get_models(model_id = "", keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{model_id: Either a model id to retrieve information for that specific model
or an empty string to retrieve information on all model associated
with the user's aWhere API account (string - optional)}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
- data.frame containing information about requested model(s)
}
\description{
\code{get_models} Calls Models endpoint of API
}
\details{
This API provides the list of available models in the aWhere platform.
Today a variety of crop growth stage models are generally available, with many more on the way.

Each models applies to particular crops from the Crops API, and after an initial
review of the available models, you can often save the model ID and simply reference the details and results as needed.
}
\examples{
\dontrun{get_models('BarleyGenericMSU')
         get_models()
}
}
\references{
https://developer.awhere.com/api/reference/models/get-models
}
