% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_planting}
\alias{get_planting}
\title{Get Planting}
\usage{
get_planting(field_id = "", planting_id = "", current = F)
}
\arguments{
\item{-}{field_id: a field ID to look within (string)}

\item{-}{planting_id: a planting ID to look for (string)}

\item{-}{current: whether to just get current plantings(T) or include historical plantings(F).
To get most recent planting record for a field, set current to TRUE and do not pass in a planting_id (boolean)}
}
\value{
- data: data.table containing information about requested field(s)
}
\description{
\code{get_planting} Gets a planting or list of plantings in the aWhere platform for which you can request weather
}
\details{
Fields are how you manage the locations for which you're tracking weather, agronomics,
models, and progress over growing seasons. By registering a field, you create a quick way
to consistently reference locations across all our APIs, and allow our modeling APIs
to better operate and tune to the conditions and status of each specific field. A Planting
is the record of a crop's season in a given field, and is where you tell the platform
about what is planted there and when it was planted.

Plantings are the way to manage crop and field activity in the aWhere API. Use this
API to record the type of crop, planting date, projections, and actuals to get the
most out of our more advanced APIs. Over time, this API also enables historical metrics
for any given field in the platform. In this function by setting an Id you can retrieve the weather
and agronomics for that location in all the other APIs.
}
\examples{
GetField(1234)
}
\references{
http://developer.awhere.com/api/reference/plantings/get-plantings
}

