% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-token.R
\name{get_token}
\alias{get_token}
\title{Get Access Token}
\usage{
get_token(uid, secret, use_environment = TRUE, apiAddress = "api.awhere.com")
}
\arguments{
\item{-}{apiAddress: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
List with three elements:#'
error: logical indicating whether there was an error
error_message: \code{NULL} if error is \code{FALSE}, a character error message otherwise
token: aWhere API access token value
}
\description{
\code{get_token} gets Access Token for V2 aWhere API
}
\details{
This script provides an aWhere access token for the current session of the API. Information for the key and secret in this
function can be found on a user's account at developer.awhere.com, under the apps.
}
\examples{
\dontrun{get_token("uid", "secret")}
\dontrun{token_response <- get_token('uid', 'secret', use_environment = FALSE)}
}
