% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccessToken.R
\name{GetAccessToken}
\alias{GetAccessToken}
\title{getAccessToken}
\usage{
GetAccessToken(uid, secret)
}
\arguments{
\item{-}{uid: Username associated with the aWhere API}

\item{-}{secret: Alphanumeric key associated with your aWhere API account}
}
\value{
None
}
\description{
\code{getAccessToken} gets Access Token for V2 aWhere API
}
\details{
this script gets the aWHERE access token for the current session of the API
}
\examples{
GetAccessToken(uid,secret)
}

