% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-functions.R
\name{getGridX}
\alias{getGridX}
\title{Get aWhere GridX from longitude coordinate(s)}
\usage{
getGridX(longitude)
}
\arguments{
\item{longitude}{one more more longitude coordinates}
}
\value{
A vector of GridX values
}
\description{
Get aWhere GridX from longitude coordinate(s)
}
\examples{
getGridX(-90)

}
