% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-functions.R
\name{getGridY}
\alias{getGridY}
\title{Get aWhere GridY from latitude coordinate(s)}
\usage{
getGridY(latitude)
}
\arguments{
\item{latitude}{one or more latitudes coordinates}
}
\value{
A vector of GridY values
}
\description{
Get aWhere GridY from latitude coordinate(s)
}
\examples{
getGridY(-90)

}
