% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-token.R
\name{load_credentials}
\alias{load_credentials}
\title{Load Credentials.}
\usage{
load_credentials(path_to_credentials)
}
\arguments{
\item{-}{path_to_credentials: absolute or relative path to the text file}
}
\value{
vector with uid and secret in positions 1, 2
}
\description{
\code{load_credentials} loads credential information from text file to use in
 calls to aWhere API
}
\details{
This script creates loads a valid aWhere API token into memory to be used in making API calls.
Instead of typing in the password and username manually and calling the get_token(uid,secret) function,
you can store this information in a txt file in some arbitrary directory.
The first line of the text file should be the provided uid/username, the
second line should be the associated secret.  A blank 3rd line should be inserted to prevent
an error from being returned by R
}
\examples{
\dontrun{load_credentials("C:/aWhere/credentials/credentials.txt")}

}
