% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_APICalls.R
\name{plan_APICalls}
\alias{plan_APICalls}
\title{Plan Logic of API Calls}
\usage{
plan_APICalls(day_start, day_end, numObsReturned, includesLeapYear)
}
\arguments{
\item{-}{JSON object returned from aWhere API}
}
\value{
list(allDates,loops)
}
\description{
\code{plan_APICalls} Calculate logic of API calls
}
\details{
Calculate logic of API calls based on the number of days that can be returned
with each call as well as start and end dates
}
