% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanupAPIReturns.R,
%   R/removeUnnecessaryColumns.R
\name{recalculateAccumulations}
\alias{recalculateAccumulations}
\title{Removes unnecessary columns from API Return}
\usage{
recalculateAccumulations(dataList)

recalculateAccumulations(dataList)
}
\arguments{
\item{-}{dataList: list of data.table to have accumulations corrected}

\item{-}{dataList: list of data.table to have accumulations corrected}
}
\description{
\code{recalculateAccumulations} manually recalculate accumulations over API calls

\code{recalculateAccumulations} manually recalculate accumulations over API calls
}
