% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanupAPIReturns.R
\name{removeUnnecessaryColumns}
\alias{removeUnnecessaryColumns}
\title{Removes unnecessary columns from API Return}
\usage{
removeUnnecessaryColumns(data, returnOnlySoilVars = FALSE)
}
\arguments{
\item{-}{data: data.table to have columns removed from}
}
\description{
\code{removeUnnecessaryColumns} Removes unnecessary columns from API Return
}
