% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_field}
\alias{update_field}
\title{Update Field}
\usage{
update_field(field_id, variable_search, value_search, variable_update,
  value_update)
}
\arguments{
\item{-}{field_id: the variable that will be used to search for the particular}

\item{-}{variable_search: the variable that will be used to search for the particular
locations that need information updated (character string).
Either "farmId" or "name"}

\item{-}{value_search: the value for the variable_search (character string)}

\item{-}{variable_update: the variable that needs to be updated (character string).
Either "farmId" or "name"}

\item{-}{value_update: the new value that will be placed (character string)}
}
\value{
- A message confirming the changes have been made
}
\description{
\code{update_field} To update details (Farm ID or FieldName) of a particular location in the aWhere API.
}
\details{
Fields are how you manage the locations for which you're tracking weather, agronomics,
models, and progress over growing seasons. By registering a field, you create a quick way
to consistently reference locations across all our APIs, and allow our modeling APIs
to better operate and tune to the conditions and status of each specific field. A Planting
is the record of a crop's season in a given field, and is where you tell the platform
about what is planted there and when it was planted.

Occasionally, you may need to update a field location detail.
Only the Farm ID and Field Name can be updated at this time.
IMPORTANT - to look for a set of data points which need update, you can only update parameters
            one at a time:
                   - only farmId
                   - only name
}
\examples{
update_field( field_id = 'field123',
				variable_search = 'farmId', value_search = 'farmA',
             variable_update = 'farmId', value_update = 'This is my territory')
}
\references{
http://developer.awhere.com/api/reference/fields/update-field
}

