% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_planting}
\alias{update_planting}
\title{Update Planting}
\usage{
update_planting(
  field_id,
  planting_id,
  planting_date = "",
  proj_yield_amount = "",
  proj_yield_units = "",
  proj_harvest_date = "",
  yield_amount = "",
  yield_units = "",
  harvest_date = "",
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token,
  apiAddressToUse = awhereEnv75247$apiAddress
)
}
\arguments{
\item{-}{apiAddressToUse: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
- A message confirming the changes have been made
}
\description{
\code{update_planting} To update details of a particular planting
}
\details{
Occasionally you will need to update a planting, changing the projections
or recording the end-of-season information for historical tracking and model tuning.
This API supports both partial and complete updates of plantings.  When updating an
entire planting, the whole object is replaced in the database. Any properties that
were previously set, and now are not, will be null'd. The required properties must
also be set even if they are changed.
}
\examples{
\dontrun{update_planting("field_test", "156036", harvest_date = "2016-02-01", yield_amount = "60", yield_units = "Bushels")}

}
\references{
https://docs.awhere.com/knowledge-base-docs/update-all-or-part-of-a-planting-in-a-field/
}
