% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateField.R
\name{updateField}
\alias{updateField}
\title{updateField}
\usage{
updateField(variableToSearch, valueToSearch, variableToChange, valueToChange)
}
\arguments{
\item{-}{variableToSearch: the variable that will be used to search for the particular
locations that need information updated (character string).
Either "id" or "farmId" or "name"}

\item{-}{valueToSearch: the value for the variableToSearch (character string)}

\item{-}{variableToUpdate: the variable that needs to be updated (character string).
Either "id" or "farmId" or "name"}

\item{-}{valueToUpdate: the new value that will be placed (character string)}
}
\value{
- A message confirming the changes have been made
}
\description{
\code{updateField} To update details (Farm ID or FieldName) of a particular location in the aWhere API.
}
\details{
Fields are how you manage the locations for which you're tracking weather, agronomics,
models, and progress over growing seasons. By registering a field, you create a quick way
to consistently reference locations across all our APIs, and allow our modeling APIs
to better operate and tune to the conditions and status of each specific field. A Planting
is the record of a crop's season in a given field, and is where you tell the platform
about what is planted there and when it was planted.

Occasionally, you may need to update a field location detail.
Only the Farm ID and Field Name can be updated at this time.
IMPORTANT - to look for a set of data points which need update, you can only update parameters
            one at a time:
                   - only fieldId
                   - only farmId
                   - only name
}
\examples{
updateField( variableToSearch = 'farmId', valueToSearch = 'farmA',
             variableToChange = 'farmId', valueToChange = 'This is my territory')
}
\references{
http://developer.awhere.com/api/reference/fields/update-field
}

