% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify-api-calls.R
\name{verify_api_calls}
\alias{verify_api_calls}
\title{Verify User Wants Make API Calls}
\usage{
verify_api_calls(grid, bypassNumCallCheck)
}
\arguments{
\item{-}{grid: data frame returned from create_awhere_grid}
}
\description{
\code{verify_api_calls} Tells user how many api calls will be made to pull the data requested
}
