% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-norms.R
\name{weather_norms_area}
\alias{weather_norms_area}
\title{weather_norms_area}
\usage{
weather_norms_area(
  polygon,
  monthday_start,
  monthday_end,
  year_start,
  year_end,
  propertiesToInclude = "",
  exclude_years = NULL,
  includeFeb29thData = TRUE,
  numcores = 2,
  bypassNumCallCheck = FALSE,
  returnSpatialData = FALSE,
  verbose = TRUE,
  maxTryCount = 3,
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token,
  apiAddressToUse = awhereEnv75247$apiAddress
)
}
\arguments{
\item{-}{apiAddressToUse: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
data.frame of requested data for dates requested
}
\description{
\code{weather_norms_area} pulls long term norm weather data from aWhere's API based on a data.frame of lat/lon, polygon or extentt
}
\details{
This function allows you to calculate the averages for weather attributes
across any range of years for which data are available.  The data pulled includes
meanTemp, maxTemp, minTemp, precipitation average, solar radiation average,
minHumidity, maxHumidity, maxWind and averageWind, along with the standard deviations
for these variables.  The polygon should be either a SpatialPolygons object or
a well-known text character string or an extent.

The data returned in this function
allow you to compare this year or previous years to the long-term normals, calculated as
the average of those weather conditions on that day in that location over the years specified.

Note about dates: The system does not adjust for any difference in dates between the location of the user
          and where data is being requested from.  It is the responsibility of the user to ensure a valid
          date range is specified given any differences in timezone.  These differences can have implications
          for whether a given date should be requested from the daily_observed functions or the forecast functions.
          Furthermore, because this function can take as input locations that may be in different timezones, it is
          the responsibility of the user to either ensure that the date range specified is valid for all relevant
          locations or to break the query into pieces.
}
\examples{
\dontrun{weather_norms_area(polygon = raster::getData('GADM', country = "Gambia", level = 0, download = T)
                              ,monthday_start = '02-01'
                              ,monthday_end = '03-10'
                              ,year_start = 2008
                              ,year_end = 2015
                              ,exclude_years =  c(2010,2011)
                              ,numcores = 2)}
}
\references{
https://docs.awhere.com/knowledge-base-docs/historical-weather-norms-by-geolocation/
}
