% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-norms.R
\name{weather_norms_fields}
\alias{weather_norms_fields}
\title{weather_norms_fields}
\usage{
weather_norms_fields(
  field_id,
  monthday_start,
  monthday_end,
  year_start,
  year_end,
  propertiesToInclude = "",
  exclude_years = NULL,
  includeFeb29thData = TRUE,
  keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token,
  apiAddressToUse = awhereEnv75247$apiAddress
)
}
\arguments{
\item{-}{apiAddressToUse: Address of aWhere API to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
dataframe of requested data for dates requested
}
\description{
\code{weather_norms_fields} pulls long term norm weather data from aWhere's API based on field id
}
\details{
This function allows you to calculate the averages for weather attributes
across any range of years for which data are available.  The data pulled includes
meanTemp, maxTemp, minTemp, precipitation average, solar radiation average,
minHumidity, maxHumidity, maxWind and averageWind, along with the standard deviations
for these variables.  The data pulled is for the field id identified.

The data returned in this function
allow you to compare this year or previous years to the long-term normals, calculated as
the average of those weather conditions on that day in that location over the years specified.

Note about dates: The system does not adjust for any difference in dates between the location of the user
          and where data is being requested from.  It is the responsibility of the user to ensure a valid
          date range is specified given any differences in timezone.  These differences can have implications
          for whether a given date should be requested from the daily_observed functions or the forecast functions
}
\examples{
\dontrun{weather_norms_fields(field_id = "field_test"
                              ,monthday_start = "06-01"
                              ,monthday_end = "09-01"
                              ,year_start = 2006
                              ,year_end = 2015)}
}
\references{
https://docs.awhere.com/knowledge-base-docs/historical-weather-norms/
}
