% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agronomic-values.R
\name{agronomic_values_fields}
\alias{agronomic_values_fields}
\title{Get Agronomic Values Fields.}
\usage{
agronomic_values_fields(field_id, day_start = "", day_end = "",
  accumulation_start_date = "", gdd_method = "standard",
  gdd_base_temp = "10", gdd_min_boundary = "10", gdd_max_boundary = "30")
}
\arguments{
\item{-}{field_id: the field_id having previously been created with the createField Function}

\item{-}{day_start: character string of start date in form: YYYY-MM-DD
Defaults to using the associated planting date if no date set}

\item{-}{day_end: character string of end date in form: YYYY-MM-DD
If Not included will return data only for start date}

\item{-}{accumulation_start_date: If you want to start counting accumulations from
before the specified start date (or before the
planting date if using the most recent Planting),
use this parameter to specify the date from which
you wish to start counting. The daily values object
will still only return the days between the start
and end date. This date must come before the start date.}

\item{-}{gdd_method: There are variety of equations available for calculating growing degree-days.
Valid entries are: 'standard', 'modifiedstandard', 'min-temp-cap', 'min-temp-constant'
See the API documentation for a description of each method.  The standard
method will be used if none is specified}

\item{-}{gdd_base_temp: The base temp to use for the any of the GDD equations. The default value of 10 will
be used if none is specified}

\item{-}{gdd_min_boundary: The minimum boundary to use in the selected GDD equation.
The behavior of this value is different depending on the equation you're using
The default value of 10 will be used if none is specified}

\item{-}{gdd_max_boundary: The max boundary to use in the selected GDD equation. The
behavior of this value is different depending on the equation you're using.
The default value of 30 will be used if none is specified}
}
\value{
data.table of requested data for dates requested
}
\description{
\code{agronomic_values_fields} calls Agronomic Values and Accumulations Endpoint of API using Field Location Construct
}
\details{
Agronomic Values are calculated numbers that can be used to show the agronomic status of a field or crop.
These figures can be used, for example, to track and predict plant growth or identify water stress.
Accumulated values allow growers to easily identify how the weather has been over the season.
Both sets of data are commonly used on small and large farms alike.  This is a very flexible API
that supports a wide variety of configurations to get exactly the data you want as efficiently as
possible. It's also designed to work with the Fields and Plantings system to reduce the amount of input.
While a planting is not required to use this API, be sure to create Plantings for your Fields in order
to get the most out of the aWhere platform.  This function returns GDDs, Pet, P/Pet, accumulated GDD,
accumulated Precipitation, accumulated Pet, accumulated P/Pet.  Returns data in API default units (metric)
}
\examples{
agronomic_values_fields('field123','2015-07-01','2015-07-31','','standard','10','10','30')
}
\references{
http://developer.awhere.com/api/reference/agronomics/values
}

