% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-token.R
\name{check_JSON}
\alias{check_JSON}
\title{Check JSON Object}
\usage{
check_JSON(jsonObject, request)
}
\arguments{
\item{-}{JSON object returned from aWhere API}
}
\value{
boolean for whether another query should be made
}
\description{
\code{check_JSON} Checks to make sure JSON object isn't an error
}
\details{
Checks that aWhere API didn't return an error code and that the token used in query hasn't expired
}
