% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkAccumulationStartDate}
\alias{checkAccumulationStartDate}
\title{Check Accumulation Start Date}
\usage{
checkAccumulationStartDate(accumulation_start_date, month_day_start)
}
\arguments{
\item{-}{accumulation_start_date: If you want to start counting accumulations from
before the specified start date (or before the
planting date if using the most recent Planting),
use this parameter to specify the date from which
you wish to start counting. The daily values object
will still only return the days between the start
and end date. This date must come before the start date.}

\item{-}{month_day_start}
}
\description{
\code{checkAccumulationStartDate} Checks to see if the accumulation start date is valid
}
