% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkCredentials}
\alias{checkCredentials}
\title{Check Valid Credentials Loaded}
\usage{
checkCredentials(keyToUse, secretToUse, tokenToUse)
}
\arguments{
\item{-}{keyToUse: aWhere API key to use}

\item{-}{secretToUse: aWhere API secret to use}

\item{-}{tokenToUse: aWhere API token to use}
}
\description{
\code{checkCredentials} Checks to see if valid aWhere API credentials are loaded
}
