% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataReturn.R
\name{checkDataReturn_norms}
\alias{checkDataReturn_norms}
\title{checkDataReturn_norms}
\usage{
checkDataReturn_norms(dataset, monthday_start, monthday_end, year_start,
  year_end, exclude_year, includeFeb29thData)
}
\arguments{
\item{-}{dataset: the data.frame to check}

\item{-}{monthday_start: character string of the first month and day for which you want to retrieve data,
in the form: MM-DD.  This is the start of your date range. e.g. '07-01' (July 1)}

\item{-}{monthday_end: character string of the last month and day for which you want to retrieve data,
in the form: MM-DD.  This is the end of your date range. e.g. '07-01' (July 1)}

\item{-}{year_start: character string of the starting year (inclusive) of the range of years for which
you're calculating norms, in the form YYYY. e.g., 2008}

\item{-}{year_end: character string of the last year (inclusive) of the range of years for which
you're calculating norms, in the form YYYY. e.g., 2015}

\item{-}{exclude_year: Year or years which you'd like to exclude from
 your range of years on which to calculate norms. To exclude
 multiple years, provide a vector of years. You must include
at least three years of data with which to calculate the norms.}

\item{-}{includeFeb29thData: Whether to keep data from Feb 29th on leap years.}
}
\description{
\code{checkDataReturn_norms} Checks to see if data return is complete from norms API
}
