% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidParamsFns.R
\name{checkValidField}
\alias{checkValidField}
\title{Check Valid Field ID}
\usage{
checkValidField(field_id, keyToUse, secretToUse, tokenToUse)
}
\arguments{
\item{-}{field_id: the field_id having previously been created with the createField Function}

\item{-}{keyToUse: aWhere API key to use}

\item{-}{secretToUse: aWhere API secret to use}

\item{-}{tokenToUse: aWhere API token to use}
}
\description{
\code{checkValidField} Checks to see if the Field ID is valid
}
