% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-daily.R
\name{daily_observed_fields}
\alias{daily_observed_fields}
\title{daily_observed_fields}
\usage{
daily_observed_fields(field_id, day_start, day_end,
  keyToUse = awhereEnv75247$uid, secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{field_id: the field_id associated with the location for which you want to pull data.
Field IDs are created using the create_field function.(string)}

\item{-}{day_start: character string of the first day for which you want to retrieve data, in the form: YYYY-MM-DD}

\item{-}{day_end: character string of the last day for which you want to retrieve data, in form: YYYY-MM-DD}

\item{-}{keyToUse: aWhere API key to use.  DO NOT USE OPTION UNLESS YOU KNOW WHAT YOU ARE DOING (optional)}

\item{-}{secretToUse: aWhere API secret to use.  DO NOT USE OPTION UNLESS YOU KNOW WHAT YOU ARE DOING (optional)}

\item{-}{tokenToUse: aWhere API token to use.  DO NOT USE OPTION UNLESS YOU KNOW WHAT YOU ARE DOING (optional)}
}
\value{
data.frame of requested data for dates requested
}
\description{
\code{daily_observed_fields} pulls historical weather data from aWhere's API based on field id
}
\details{
This function returns weather data on Min/Max Temperature, Precipitation,
Min/Max Humidity, Solar Radiation, and Maximum Wind Speed,
Morning Max Windspeed, and Average Windspeed for the field id specified. 
Default units are returned by the API. 

The Weather APIs provide access to aWhere's agriculture-specific Weather Terrain system,
and allows retrieval and integration of data across all different time ranges, long term normals,
daily observed, current weather, and forecasts. These APIs are designed for efficiency,
allowing you to customize the responses to return just the attributes you need.

Understanding the recent and long-term daily weather is critical for making in-season decisions.
This API opens the weather attributes that matter most to agriculture.
}
\examples{
\dontrun{daily_observed_fields(field_id = 'field_test',day_start = '2016-04-28',day_end = '2017-05-01')}
}
\references{
http://developer.awhere.com/api/reference/weather/observations
}
