% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-daily.R
\name{daily_observed_latlng}
\alias{daily_observed_latlng}
\title{daily_observed_latlng}
\usage{
daily_observed_latlng(latitude, longitude, day_start, day_end,
  keyToUse = awhereEnv75247$uid, secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{latitude: the latitude of the requested location (double)}

\item{-}{longitude: the longitude of the requested locations (double)}

\item{-}{day_start: character string of the first day for which you want to retrieve data, in the form: YYYY-MM-DD}

\item{-}{day_end: character string of the last day for which you want to retrieve data, in the form: YYYY-MM-DD}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
data.frame of requested data for dates requested
}
\description{
\code{daily_observed_latlng} pulls historical weather data from aWhere's API based on latitude & longitude
}
\details{
This function returns weather data on Min/Max Temperature, Precipitation,
Min/Max Humidity, Solar Radiation, and Maximum Wind Speed,
Morning Max Windspeed, and Average Windspeed for the location specified by latitude and longitude. 
Default units are returned by the API. Latitude and longitude must be in decimal degrees.

The Weather APIs provide access to aWhere's agriculture-specific Weather Terrain system,
and allows retrieval and integration of data across all different time ranges, long term normals,
daily observed, current weather, and forecasts. These APIs are designed for efficiency,
allowing you to customize the responses to return just the attributes you need.

Understanding the recent and long-term daily weather is critical for making in-season decisions.
This API opens the weather attributes that matter most to agriculture.
}
\examples{
\dontrun{daily_observed_latlng(latitude = 39.8282, longitude = -98.5795,day_start = '2014-04-28',day_end = '2015-05-01')}
}
\references{
http://developer.awhere.com/api/reference/weather/observations/geolocation
}
