% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-token.R
\name{get_token}
\alias{get_token}
\title{Get Access Token}
\usage{
get_token(uid, secret)
}
\arguments{
\item{-}{uid: Consumer key associated with the user's aWhere API account}

\item{-}{secret: Consumer secret associated the user's aWhere API account}
}
\value{
None
}
\description{
\code{get_token} gets Access Token for V2 aWhere API
}
\details{
This script provides an aWhere access token for the current session of the API. Information for the key and secret in this 
function can be found on a user's account at developer.awhere.com, under the apps.
}
\examples{
\dontrun{get_token("uid", "secret")}
}

