% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-token.R
\name{get_token}
\alias{get_token}
\title{Get Access Token}
\usage{
get_token(uid, secret)
}
\arguments{
\item{-}{uid: Username associated with the aWhere API}

\item{-}{secret: Alphanumeric key associated with your aWhere API account}
}
\value{
None
}
\description{
\code{get_token} gets Access Token for V2 aWhere API
}
\details{
this script gets the aWHERE access token for the current session of the API
}
\examples{
\dontrun{get_token(uid, secret)}
}

