% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-token.R
\name{get_token}
\alias{get_token}
\title{Get Access Token}
\usage{
get_token(uid, secret, use_environment = TRUE)
}
\arguments{
\item{-}{uid: Consumer key associated with the user's aWhere API account}

\item{-}{secret: Consumer secret associated the user's aWhere API account}

\item{-}{use_enviroment: Optional logical value, determines whether API access
token will be saved in a local locked environment in addition to being returned
by the function. Defaults to \code{TRUE} to avoid breaking existing code.}
}
\value{
List with three elements:#'
error: logical indicating whether there was an error
error_message: \code{NULL} if error is \code{FALSE}, a character error message otherwise
token: aWhere API access token value
}
\description{
\code{get_token} gets Access Token for V2 aWhere API
}
\details{
This script provides an aWhere access token for the current session of the API. Information for the key and secret in this
function can be found on a user's account at developer.awhere.com, under the apps.
}
\examples{
\dontrun{get_token("uid", "secret")}
\dontrun{token_response <- get_token('uid', 'secret', use_environment = FALSE)}
}
