% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-norms.R
\name{weather_norms_area}
\alias{weather_norms_area}
\title{weather_norms_area}
\usage{
weather_norms_area(polygon, monthday_start, monthday_end, year_start,
  year_end, propertiesToInclude = "", exclude_years = NULL,
  includeFeb29thData = TRUE, numcores = 2,
  bypassNumCallCheck = FALSE, keyToUse = awhereEnv75247$uid,
  secretToUse = awhereEnv75247$secret,
  tokenToUse = awhereEnv75247$token)
}
\arguments{
\item{-}{polygon: either a SpatialPolygons object, well-known text string, or extent from raster package}

\item{-}{monthday_start: character string of the first month and day for which you want to retrieve data,
in the form: MM-DD.  This is the start of your date range. e.g. '07-01' (July 1) (required)}

\item{-}{monthday_end: character string of the last month and day for which you want to retrieve data,
in the form: MM-DD.  This is the end of your date range. e.g. '07-01' (July 1) (required)}

\item{-}{year_start: character string of the starting year (inclusive) of the range of years for which
you're calculating norms, in the form YYYY. e.g., 2008 (required)}

\item{-}{year_end: character string of the last year (inclusive) of the range of years for which
you're calculating norms, in the form YYYY. e.g., 2015 (required)}

\item{-}{propertiesToInclude: character vector of properties to retrieve from API.  Valid values are meanTemp, maxTemp, minTemp, precipitation, solar, maxHumidity, minHumidity, dailyMaxWind (optional)}

\item{-}{exclude_year: Year or years which you'd like to exclude from
 your range of years on which to calculate norms. To exclude
 multiple years, provide a vector of years. You must include
at least three years of data with which to calculate the norms. (numeric, optional)}

\item{-}{includeFeb29thData: Whether to keep data from Feb 29th on leap years.  Because weather/agronomics
summary statistics are calculated via the calendar date and 3 years are required
to generate a value, data from this date is more likely to be NA.  ALlows user
to drop this data to avoid later problems (defaults to TRUE)}

\item{-}{numcores: number of cores to use in parallel loop. To check number of available cores: parallel::detectCores().
If you receive an error regarding the speed you are making calls, reduce this number}

\item{-}{bypassNumCallCheck: set to TRUE to avoid prompting the user to confirm that they want to begin making API calls}

\item{-}{keyToUse: aWhere API key to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{secretToUse: aWhere API secret to use.  For advanced use only.  Most users will not need to use this parameter (optional)}

\item{-}{tokenToUse: aWhere API token to use.  For advanced use only.  Most users will not need to use this parameter (optional)}
}
\value{
data.frame of requested data for dates requested
}
\description{
\code{weather_norms_area} pulls long term norm weather data from aWhere's API based on spatial polygon or extent
}
\details{
This function allows you to calculate the averages for weather attributes
across any range of years for which data are available.  The data pulled includes
meanTemp, maxTemp, minTemp, precipitation average, solar radiation average,
minHumidity, maxHumidity, maxWind and averageWind, along with the standard deviations
for these variables.  The polygon should be either a SpatialPolygons object or
a well-known text character string or an extent.

The data returned in this function
allow you to compare this year or previous years to the long-term normals, calculated as
the average of those weather conditions on that day in that location over the years specified.

Note about dates: The system does not adjust for any difference in dates between the location of the user
          and where data is being requested from.  It is the responsibility of the user to ensure a valid
          date range is specified given any differences in timezone.  These differences can have implications
          for whether a given date should be requested from the daily_observed functions or the forecast functions.
          Furthermore, because this function can take as input locations that may be in different timezones, it is
          the responsibility of the user to either ensure that the date range specified is valid for all relevant
          locations or to break the query into pieces.
}
\examples{
\dontrun{weather_norms_area(polygon = raster::getData('GADM', country = "Gambia", level = 0, download = T)
                              ,monthday_start = '02-01'
                              ,monthday_end = '03-10'
                              ,year_start = 2008
                              ,year_end = 2015
                              ,exclude_years =  c(2010,2011)
                              ,numcores = 2)}
}
\references{
http://developer.awhere.com/api/reference/weather/norms
}
