% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather-norms.R
\name{weather_norms_fields}
\alias{weather_norms_fields}
\title{weather_norms_fields.}
\usage{
weather_norms_fields(field_id, monthday_start = "", monthday_end = "",
  year_start = "", year_end = "", exclude_years = "")
}
\arguments{
\item{-}{field_id: the field_id having previously been created with the createField Function}

\item{-}{monthday_start: character string of the month and day for the start
of the range of days you are calculating norms for, e.g., '07-01' (July 1)}

\item{-}{monthday_end: character string of the month and day for the end of the
range of days you are calculating norms for, e.g., '07-10' (July 10)}

\item{-}{year_start: the starting year (inclusive) of the range of years for which
you're calculating norms, as a string, e.g., '2008'}

\item{-}{year_start: the end year (inclusive) of the range of years for which you're
calculating norms, as a string, e.g., '2015'}

\item{-}{exclude_years: You can opt to exclude one or more years from the range, and
it's values will not be included in the averages. To exclude
multiple years, separate them with a comma. Note: You must always have
at least three years of data to average}
}
\value{
data.table of requested data for dates requested
}
\description{
\code{weather_norms_fields} calls Historic Weather Norms Endpoint of API using Field Location Construct
}
\details{
This is a flexible API that allows you to calculate the averages for weather attribute
across any range of years for which we have data. Whereas the Daily Observed API only
supports up to 30 months of daily data, this API allow you to compare this year and
the previous year to the long-term normals (however many years you want to include).
Each day's worth of data also includes the standard deviation for the average.
}
\examples{
weather_norms_fields('field123','07-01', '07-10', '2008', '2015','2010,2011')
}
\references{
http://developer.awhere.com/api/reference/weather/norms
}

